Deletion Error fixed now Works
Reason of error is when making requests using jquery even if we post boolean django reads it as a string and when we deal with it as boolean it caused problem. This solution works because it uses that variable as a string but i think it isn't a great solution.